Clean up service defaults - Make it look more like the default template.
