+++ src/Ordering.API/Extensions/Extensions.cs
@@ -2,6 +2,9 @@
 {
     public static void AddApplicationServices(this IHostApplicationBuilder builder)
     {
+        // Add the authentication services to DI
+        builder.AddDefaultAuthentication();
+
         // Pooling is disabled because of the following error:
         // Unhandled exception. System.InvalidOperationException:
         // The DbContext of type 'OrderingContext' cannot be pooled because it does not have a public constructor accepting a single parameter of type DbContextOptions or has more than one constructor.
@@ -20,6 +23,7 @@
         builder.Services.AddOptions<OrderingOptions>()
                 .BindConfiguration(nameof(OrderingOptions));
 
+        builder.Services.AddHttpContextAccessor();
         builder.Services.AddTransient<IIdentityService, IdentityService>();
 
         // Configure mediatR
