+++ src/Ordering.BackgroundTasks/Ordering.BackgroundTasks.csproj
@@ -3,7 +3,6 @@
   <PropertyGroup>
     <TargetFramework>net8.0</TargetFramework>
     <EnableConfigurationBindingGenerator>true</EnableConfigurationBindingGenerator>
-    <PublishAot>true</PublishAot>
   </PropertyGroup>
 
   <ItemGroup>
