--- src/eShop.ServiceDefaults/HttpClientAuthorizationDelegatingHandler.cs
@@ -1,35 +1,0 @@
-﻿using System.Net.Http.Headers;
-using Microsoft.AspNetCore.Authentication;
-using Microsoft.AspNetCore.Http;
-
-namespace eShop.ServiceDefaults;
-
-internal class HttpClientAuthorizationDelegatingHandler : DelegatingHandler
-{
-    private readonly IHttpContextAccessor _httpContextAccessor;
-
-    public HttpClientAuthorizationDelegatingHandler(IHttpContextAccessor httpContextAccessor)
-    {
-        _httpContextAccessor = httpContextAccessor;
-    }
-
-    public HttpClientAuthorizationDelegatingHandler(IHttpContextAccessor httpContextAccessor, HttpMessageHandler innerHandler) : base(innerHandler)
-    {
-        _httpContextAccessor = httpContextAccessor;
-    }
-
-    protected override async Task<HttpResponseMessage> SendAsync(HttpRequestMessage request, CancellationToken cancellationToken)
-    {
-        if (_httpContextAccessor.HttpContext is HttpContext context)
-        {
-            var accessToken = await context.GetTokenAsync("access_token");
-
-            if (accessToken is not null)
-            {
-                request.Headers.Authorization = new AuthenticationHeaderValue("Bearer", accessToken);
-            }
-        }
-
-        return await base.SendAsync(request, cancellationToken);
-    }
-}
