+++ src/eShop.ServiceDefaults/HttpClientExtensions.cs
@@ -1,5 +1,8 @@
-﻿using Microsoft.Extensions.DependencyInjection;
+﻿using Microsoft.AspNetCore.Http;
+using System.Net.Http.Headers;
+using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection.Extensions;
+using Microsoft.AspNetCore.Authentication;
 
 namespace eShop.ServiceDefaults;
 
@@ -7,10 +10,43 @@ public static class HttpClientExtensions
 {
     public static IHttpClientBuilder AddAuthToken(this IHttpClientBuilder builder)
     {
+        builder.Services.AddHttpContextAccessor();
+
         builder.Services.TryAddTransient<HttpClientAuthorizationDelegatingHandler>();
 
         builder.AddHttpMessageHandler<HttpClientAuthorizationDelegatingHandler>();
 
         return builder;
     }
+
+    private class HttpClientAuthorizationDelegatingHandler : DelegatingHandler
+    {
+        private readonly IHttpContextAccessor _httpContextAccessor;
+
+        public HttpClientAuthorizationDelegatingHandler(IHttpContextAccessor httpContextAccessor)
+        {
+            _httpContextAccessor = httpContextAccessor;
+        }
+
+        public HttpClientAuthorizationDelegatingHandler(IHttpContextAccessor httpContextAccessor, HttpMessageHandler innerHandler) : base(innerHandler)
+        {
+            _httpContextAccessor = httpContextAccessor;
+        }
+
+        protected override async Task<HttpResponseMessage> SendAsync(HttpRequestMessage request, CancellationToken cancellationToken)
+        {
+            if (_httpContextAccessor.HttpContext is HttpContext context)
+            {
+                var accessToken = await context.GetTokenAsync("access_token");
+
+                if (accessToken is not null)
+                {
+                    request.Headers.Authorization = new AuthenticationHeaderValue("Bearer", accessToken);
+                }
+            }
+
+            return await base.SendAsync(request, cancellationToken);
+        }
+    }
+
 }
