--- src/eShop.ServiceDefaults/JsonDefaults.cs
@@ -1,11 +1,0 @@
-﻿using System.Text.Json;
-
-namespace eShop.ServiceDefaults;
-
-public static class JsonDefaults
-{
-    public static readonly JsonSerializerOptions CaseInsensitiveOptions = new()
-    {
-        PropertyNameCaseInsensitive = true
-    };
-}
